Python script to prepare a network distance matrix in QGIS
This script allows prepared data to be used in the generation of a network distance matrix in QGIS. The output is a modified (existing) shapefile with new columns appended to it. This script needs to be put in the scripts folder and can then be accessed in the Processing Toolbox.